+++ src/deps.cljs
@@ -1,1 +1,1 @@
-{:npm-deps { "@headlessui/react" "^1.4.3"}}
+{:npm-deps { "@headlessui/react" "^1.5.0"}}